WIP – move store plugin configuration to config/
