+++ lib/config/defaults.js
@@ -2,6 +2,8 @@ module.exports = getDefaults
 
 var path = require('path')
 
+var storePreAuthHook = require('./store/pre-auth-hook')
+
 function getDefaults () {
   var projectPath = process.cwd()
   var pkg = require(path.join(projectPath, 'package.json'))
@@ -18,7 +20,14 @@ function getDefaults () {
       protocol: 'http'
     },
     db: {},
+
+    // core modules
     account: {},
-    admin: {}
+    admin: {},
+    store: {
+      hooks: {
+        onPreAuth: storePreAuthHook
+      }
+    }
   }
 }
