+++ lib/config/store/pre-auth-hook.js
@@ -7,6 +7,8 @@ var Boom = require('boom')
  * are "user/{accountId}". `onStorePreAuth` checks if the bearerToken is
  * a valid session Id, and if it belongs to the right user. If either fails,
  * it responds with an `unauthorized` error
+ *
+ * See https://github.com/hoodiehq/hoodie-store-server#optionshooks
  */
 function onStorePreAuth (request, reply) {
   var server = request.connection.server
