+++ lib/hapi.js
@@ -1,24 +1,16 @@
 module.exports = hapi
 
-var onStorePreAuth = require('./core-modules-glue-code/store-pre-auth')
-
 function hapi (config, couchConfig, callback) {
   var database = require('./database')(config)
   var usersDb = database(couchConfig.authentication_db)
 
   usersDb.constructor.plugin(require('pouchdb-users'))
 
-  var storeOptions = {
-    usersDb: usersDb,
-    hooks: {
-      onPreAuth: onStorePreAuth
-    }
-  }
-
+  config.store.usersDb = usersDb
   if (config.db.url) {
-    storeOptions.couchdb = removeAuth(config.db.url)
+    config.store.couchdb = removeAuth(config.db.url)
   } else {
-    storeOptions.PouchDB = usersDb.constructor
+    config.store.PouchDB = usersDb.constructor
   }
 
   usersDb.installUsersBehavior()
@@ -51,7 +43,7 @@ function hapi (config, couchConfig, callback) {
       }
     }, {
       register: require('@hoodie/store'),
-      options: storeOptions,
+      options: config.store,
       routes: {
         prefix: '/hoodie/store/api'
       }
