+++ lib/index.js
@@ -14,8 +14,8 @@ function getHoodieServer (options, callback) {
   log.level = options.loglevel || 'warn'
 
   var config = require('./config')(options)
-
   var hapiConfig = {}
+  var willWritePouchDbFiles = !options.dbUrl && !options.inMemory
 
   if (log.level === 'debug') {
     hapiConfig = {
@@ -27,6 +27,9 @@ function getHoodieServer (options, callback) {
   }
 
   mkdirp.sync(config.paths.data)
+  if (willWritePouchDbFiles) {
+    mkdirp.sync(config.db.prefix)
+  }
 
   var publicFolderExists
   try {
